Moved hook version to its own folder and exported from index
